test: 관광지 목록 조회 테스트 코드 구현

+++ src/main/java/com/travelsketch/travel/api/controller/attraction/AttractionQueryController.java
@@ -33,7 +33,7 @@ public class AttractionQueryController {
     }
 
     @GetMapping("/{sidoId}/guguns")
-    public ApiResponse<List<GugunResponse>> searchGuguns(@PathVariable Long sidoId) {
+    public ApiResponse<List<GugunResponse>> searchGuguns(@PathVariable Integer sidoId) {
 
         List<GugunResponse> responses = areaQueryService.searchGuguns(sidoId);
